% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_version.R
\name{man_set_version}
\alias{man_set_version}
\alias{man_get_version}
\title{Access or modify the active GRASS doc version}
\usage{
man_set_version(version = 7.8)

man_get_version()
}
\arguments{
\item{version}{[Double] The desired version of GRASS (7.8, 7.9, etc...)}
}
\value{
The function `man_get_version` will return the currently set version, and
`man_set_version` will do so invisibly.
}
\description{
This set of function can be used to get the version used by the package or
to set the desired version.
}
\examples{
man_set_version(7.9)
man_get_version()

}
