% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassmodules78}
\alias{grassmodules78}
\title{GRASS 7.8 modules names and descriptions}
\format{
A data frame with 528 rows and 2 columns.
\describe{
  \item{name}{character, module name}
  \item{description}{character, module description}
  \item{addon}{logical, whether the module is an addon}
}
}
\source{
\url{https://grass.osgeo.org/grass78/manuals/full_index.html}
\url{https://grass.osgeo.org/grass78/manuals/addons/}
}
\usage{
grassmodules78
}
\description{
A dataset containing the names and descriptions of all GRASS 7.8 modules
and addons, scraped from the GRASS manual full index page.
}
\keyword{datasets}
